+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Update image import path from Supabase Storage
+- Request and return all image descriptions in one shot
 
 ## [0.7.0] - 2025-08-11